requirements.txt
flask
imblearn
matplotlib
numpy
pandas
psycopg2-binary
scikit-learn
seaborn
torch
transformers